Milestone 13 - FAT16 driver expansion, fixes

+++ src/common/wyrd.h
@@ -5,10 +5,10 @@ typedef short i16;
 typedef int   i32;
 typedef long  i64;
 
-typedef unsigned char  u8;
-typedef unsigned short u16;
-typedef unsigned int   u32;
-typedef unsigned long  u64;
+typedef unsigned char         u8;
+typedef unsigned short        u16;
+typedef unsigned int          u32;
+typedef unsigned long long    u64;
 
 typedef unsigned char  bool;
